Move download function into client utilities
Clients are the only users, and it avoids some extra dependencies in services.
